chore: migrate collectible components to UI, ref leather-io/issues#222
