+++ package.json
@@ -205,7 +205,6 @@
     "cross-fetch": "4.0.0",
     "css-loader": "7.1.0",
     "dayjs": "1.11.8",
-    "dompurify": "3.1.4",
     "dotenv": "16.4.5",
     "ecdsa-sig-formatter": "1.0.11",
     "formik": "2.4.5",
@@ -245,7 +244,6 @@
     "ts-debounce": "4.0.0",
     "url": "0.11.3",
     "url-join": "5.0.0",
-    "use-events": "1.4.2",
     "use-latest": "1.2.1",
     "valid-url": "1.0.9",
     "varuint-bitcoin": "1.1.2",
@@ -289,7 +287,6 @@
     "@types/argon2-browser": "1.18.4",
     "@types/bn.js": "5.1.5",
     "@types/chrome": "0.0.260",
-    "@types/dompurify": "3.0.5",
     "@types/download": "8.0.5",
     "@types/html-webpack-plugin": "3.2.9",
     "@types/jsdom": "21.1.3",
