--- src/app/components/inscription-preview-card/components/inscription-image.tsx
@@ -1,11 +1,0 @@
-interface InscriptionImageProps {
-  src: string;
-}
-export function InscriptionImage({ src }: InscriptionImageProps) {
-  return (
-    <img
-      src={src}
-      style={{ width: '100%', height: '100%', aspectRatio: '1 / 1', objectFit: 'cover' }}
-    />
-  );
-}
