+++ src/app/components/inscription-preview-card/inscription-preview.tsx
@@ -1,10 +1,8 @@
 import { BoxProps, Flex } from 'leather-styles/jsx';
 
 import type { Inscription } from '@leather.io/models';
-import { OrdinalAvatarIcon } from '@leather.io/ui';
+import { InscriptionImage, InscriptionPreviewLayout, OrdinalAvatarIcon } from '@leather.io/ui';
 
-import { InscriptionImage } from './inscription-image';
-import { InscriptionPreviewContainer } from './inscription-preview-container';
 import { InscriptionText } from './inscription-text';
 
 interface InscriptionPreviewProps extends BoxProps {
@@ -14,25 +12,25 @@ export function InscriptionPreview({ inscription, ...props }: InscriptionPreview
   switch (inscription.mimeType) {
     case 'image': {
       return (
-        <InscriptionPreviewContainer {...props}>
+        <InscriptionPreviewLayout {...props}>
           <InscriptionImage src={inscription.src} />
-        </InscriptionPreviewContainer>
+        </InscriptionPreviewLayout>
       );
     }
     case 'text': {
       return (
-        <InscriptionPreviewContainer {...props}>
+        <InscriptionPreviewLayout {...props}>
           <InscriptionText contentSrc={inscription.src} />
-        </InscriptionPreviewContainer>
+        </InscriptionPreviewLayout>
       );
     }
     case 'other': {
       return (
-        <InscriptionPreviewContainer {...props}>
+        <InscriptionPreviewLayout {...props}>
           <Flex alignItems="center" height="100%" justifyContent="center">
             <OrdinalAvatarIcon size="xl" />
           </Flex>
-        </InscriptionPreviewContainer>
+        </InscriptionPreviewLayout>
       );
     }
     default:
