+++ src/app/features/bitcoin-choose-fee/bitcoin-choose-fee.tsx
@@ -3,6 +3,7 @@ import { useState } from 'react';
 import { Box, Flex, FlexProps, Stack } from 'leather-styles/jsx';
 
 import type { BtcFeeType, Money } from '@leather.io/models';
+import { LoadingSpinner } from '@leather.io/ui';
 import { formatMoney } from '@leather.io/utils';
 
 import type { TransferRecipient } from '@shared/models/form.model';
@@ -11,7 +12,6 @@ import { BitcoinCustomFee } from '@app/components/bitcoin-custom-fee/bitcoin-cus
 import { MAX_FEE_RATE_MULTIPLIER } from '@app/components/bitcoin-custom-fee/hooks/use-bitcoin-custom-fee';
 import { OnChooseFeeArgs } from '@app/components/bitcoin-fees-list/bitcoin-fees-list';
 import { AvailableBalance, Card } from '@app/components/layout';
-import { LoadingSpinner } from '@app/components/loading-spinner';
 import { useCurrentBtcCryptoAssetBalanceNativeSegwit } from '@app/query/bitcoin/balance/btc-balance-native-segwit.hooks';
 import { useIsPrivateMode } from '@app/store/settings/settings.selectors';
