--- src/app/features/collectibles/components/image-unavailable.tsx
@@ -1,16 +1,0 @@
-import { styled } from 'leather-styles/jsx';
-
-import { Eye1ClosedIcon } from '@leather.io/ui';
-
-import { CollectiblePlaceholderLayout } from './_collectible-types/collectible-placeholder.layout';
-
-export function ImageUnavailable() {
-  return (
-    <CollectiblePlaceholderLayout>
-      <Eye1ClosedIcon />
-      <styled.span pt="space.02" px="space.04" textStyle="label.03">
-        Image currently unavailable
-      </styled.span>
-    </CollectiblePlaceholderLayout>
-  );
-}
