+++ src/app/features/collectibles/components/stacks/stacks-bns-name.tsx
@@ -1,8 +1,7 @@
 import StacksNftBns from '@assets/images/stacks-nft-bns.png';
+import { SendCryptoAssetSelectors } from '@tests/selectors/send.selectors';
 
-import { StxAvatarIcon } from '@leather.io/ui';
-
-import { CollectibleItemLayout } from '../collectible-item.layout';
+import { CollectibleItemLayout, StxAvatarIcon } from '@leather.io/ui';
 
 export function StacksBnsName(props: { bnsName: string }) {
   const { bnsName } = props;
@@ -12,6 +11,7 @@ export function StacksBnsName(props: { bnsName: string }) {
       collectibleTypeIcon={<StxAvatarIcon size="lg" />}
       subtitle="Bitcoin Naming System"
       title={bnsName}
+      testId={SendCryptoAssetSelectors.InscriptionSendButton}
     >
       <img alt="nft image" src={StacksNftBns} width="100px" />
     </CollectibleItemLayout>
